+++ src/lazy_object_proxy/utils.py
@@ -1,3 +1,15 @@
+ # flake8: noqa
+try:
+    from .utils_py3 import __aenter__
+    from .utils_py3 import __aexit__
+    from .utils_py3 import __aiter__
+    from .utils_py3 import __anext__
+    from .utils_py3 import __await__
+    from .utils_py3 import await_
+except (ImportError, SyntaxError):
+    await_ = None
+
+
 def identity(obj):
     return obj
 
@@ -11,26 +23,3 @@ class cached_property(object):
             return self
         value = obj.__dict__[self.func.__name__] = self.func(obj)
         return value
-
-
-try:
-    exec("""
-from inspect import isawaitable
-
-
-async def do_await(obj):
-    return await obj
-
-
-def do_yield_from(gen):
-    return (yield from gen)
-
-
-def await_(obj):
-    if isawaitable(obj):
-        return do_await(obj).__await__()
-    else:
-        return do_yield_from(obj)
-""")
-except (ImportError, SyntaxError):
-    await_ = None